Remove server_test.go
